# api-specs/swagger.json
{
  "info": {
    "title": "Application APIs",
    "version": "1.0.0",
    "description": "Application APIs"
  },
  "basePath": "/",
  "swagger": "2.0",
  "paths": {
...
    "/departament": {
      "get": {
        "description": "get Departament",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get Departament",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Departamentresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save Departament",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Departament",
            "schema": {
              "type": "object",
              "properties": {
                "Departament": {
                  "$ref": "#/definitions/Departament"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save Departament",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Departamentresponse"
                }
              }
            }
          }
        }
      }
    },
    "/departament/{id}": {
      "get": {
        "description": "get Departament with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Departament ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get Departament",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Departamentresponse"
                }
              }
            }
          },
          "404": {
            "description": "Departament not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove Departament with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Departament ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete Departament",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Departamentresponse"
                }
              }
            }
          },
          "404": {
            "description": "Departament not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update Departament with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Departament ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "Departament",
            "schema": {
              "type": "object",
              "properties": {
                "Departament": {
                  "$ref": "#/definitions/Departament"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Departament",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Departamentresponse"
                }
              }
            }
          },
          "404": {
            "description": "Departament not found",
            "content": {}
          }
        }
      }
    },
    "/education-base": {
      "get": {
        "description": "get EducationBase",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get EducationBase",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/EducationBaseresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save EducationBase",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "EducationBase",
            "schema": {
              "type": "object",
              "properties": {
                "EducationBase": {
                  "$ref": "#/definitions/EducationBase"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save EducationBase",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/EducationBaseresponse"
                }
              }
            }
          }
        }
      }
    },
    "/education-base/{id}": {
      "get": {
        "description": "get EducationBase with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "EducationBase ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get EducationBase",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/EducationBaseresponse"
                }
              }
            }
          },
          "404": {
            "description": "EducationBase not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove EducationBase with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "EducationBase ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete EducationBase",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/EducationBaseresponse"
                }
              }
            }
          },
          "404": {
            "description": "EducationBase not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update EducationBase with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "EducationBase ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "EducationBase",
            "schema": {
              "type": "object",
              "properties": {
                "EducationBase": {
                  "$ref": "#/definitions/EducationBase"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update EducationBase",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/EducationBaseresponse"
                }
              }
            }
          },
          "404": {
            "description": "EducationBase not found",
            "content": {}
          }
        }
      }
    },
    "/group": {
      "get": {
        "description": "get Group",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get Group",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Groupresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save Group",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Group",
            "schema": {
              "type": "object",
              "properties": {
                "Group": {
                  "$ref": "#/definitions/Group"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save Group",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Groupresponse"
                }
              }
            }
          }
        }
      }
    },
    "/group/{id}": {
      "get": {
        "description": "get Group with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get Group",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Groupresponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove Group with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete Group",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Groupresponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update Group with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "Group",
            "schema": {
              "type": "object",
              "properties": {
                "Group": {
                  "$ref": "#/definitions/Group"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Group",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Groupresponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Departament": {
      "properties": {
        "name": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Departamentresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "EducationBase": {
      "properties": {
        "name": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "EducationBaseresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Group": {
      "properties": {
        "name": {
          "type": "string"
        },
        "courseId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Groupresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "courseId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}